requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy[asyncio]>=2.0.30
aiosqlite>=0.19.0
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
httpx>=0.27.0
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
